refactor fastai artifact and pin fastai dependency version
